condensed send test for faster testing
