+++ components/Send/components/Confirm.tsx
@@ -87,9 +87,13 @@ const Confirm: React.FunctionComponent<ConfirmProps> = ({ closeModal, confirmSen
                 />
               </View>
               <FadeText style={{ marginTop: 10 }}>{translate('send.confirm-memo') as string}</FadeText>
-              <RegText testID="send.confirm-memo">{`${to.memo || ''}${to.memo ? '\n' : ''}${
-                to.includeUAMemo ? 'Reply to: \n' + uaAddress : ''
-              }`}</RegText>
+              <RegText
+                testID="send.confirm-memo"
+                >
+                {`${to.memo || ''}${to.memo ? '\n' : ''}${
+                  to.includeUAMemo ? 'Reply to: \n' + uaAddress : ''
+                }`}
+              </RegText>
             </View>
           );
         })}
